requirements.txt
asyncio==3.4.3
sanic==0.5.4
uvloop==0.8.0
pyzmq=16.0.3